feat[#13]: Create authentication class and tests
