+++ MarvelApp/MarvelApp/Network/Entity/Auth.swift
