+++ MarvelApp/MarvelApp/Network/Entity/NetworkServiceError.swift
